piechart legend

+++ src/app/accounts/signup/signup.component.ts
@@ -37,24 +37,26 @@ export class SignupComponent {
     if (this.signupForm.invalid) return;
 
     const formData = this.signupForm.value as SignupI;
-    console.log(formData);
 
     // Retrieve the existing data from localStorage
     const existingData = localStorage.getItem('signupDetails');
+    console.log(existingData)
 
     // Initialize an array to hold the data (existing + new)
     let updatedData = [];
 
     const password = this.signupForm.get('password')?.value;
     const confirmPassword = this.signupForm.get('confirmPassword')?.value;
+    const email = this.signupForm.get('email')?.value;
 
     if (password === confirmPassword) {
-      console.log('valid');
 
       if (existingData) {
+       
         try {
           // Parse the existing data if it exists and is valid JSON
           updatedData = JSON.parse(existingData);
+          console.log(updatedData)
 
           // Ensure updatedData is an array, otherwise initialize it as an empty array
           if (!Array.isArray(updatedData)) {
@@ -64,15 +66,21 @@ export class SignupComponent {
           console.error('Error parsing localStorage data', error);
           updatedData = []; // Reset to an empty array if parsing fails
         }
+        let checkEmailExist = updatedData.filter((a) => a.email.toLowerCase() === email.toLowerCase())
+        console.log(checkEmailExist)
+
+        if (checkEmailExist.length < 1) {
+          // Add the new form data to the array
+          updatedData.push(formData);
+
+          // Store the updated array back to localStorage as a string
+          localStorage.setItem('signupDetails', JSON.stringify(updatedData));
+          alert('registration successful')
+        } else {
+          alert('Users exist')
+        }
+
       }
-
-      // Add the new form data to the array
-      updatedData.push(formData);
-
-      // Store the updated array back to localStorage as a string
-      localStorage.setItem('signupDetails', JSON.stringify(updatedData));
-      console.log(updatedData);
-      console.log(existingData);
     } else {
       this.confirmUser = true;
       console.log("We couldn't confirm your password")
